+++ src/main/kotlin/domains/Lancamento.kt
@@ -4,7 +4,7 @@ import java.util.*
 
 data class Lancamento(
     val id: Int = 0,
-    val valor: Float,
+    val valor: Double,
     val data: Date,
     val comentario: String,
     val subcategoria: SubCategoria? = null
